requirements.txt
pyyaml
requests
coveralls
hashlib
